requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
django-channels==0.7.0
django-cors-headers==4.0.0
django-rest==0.8.7
djangorestframework==3.14.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwt==1.3.1
msgpack==1.0.5
oauthlib==3.2.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==2.0.2
zope.interface==6.0
